feat: separate abi hook from write and wait
