+++ packages/react-common/src/hooks/useContractAbi.ts
@@ -1,0 +1,20 @@
+import {
+  ContractFqn,
+  ContractVersion,
+  loadContract,
+} from '@0xflair/contracts-registry';
+import { useMemo } from 'react';
+
+export type Config = {
+  contractVersion?: ContractVersion;
+  contractFqn: ContractFqn;
+};
+
+export const useContractAbi = ({ contractVersion, contractFqn }: Config) => {
+  const contract = useMemo(
+    () => loadContract(contractFqn, contractVersion),
+    [contractFqn, contractVersion]
+  );
+
+  return contract.artifact.abi;
+};
