+++ packages/react-nft-collections/src/hooks/extensions/metadata/useBaseTokenUriUpdater.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { BytesLike } from '@ethersproject/bytes';
 import { Provider } from '@ethersproject/providers';
 import { Signer } from 'ethers';
@@ -17,10 +17,14 @@ export const useBaseTokenUriUpdater = ({
   signerOrProvider,
   args,
 }: Config) => {
-  return useContractWriteAndWait({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn:
       'collections/ERC721/extensions/ERC721PrefixedMetadataExtension',
+  });
+
+  return useContractWriteAndWait({
+    contractInterface,
     contractAddress,
     signerOrProvider,
     functionName: 'setBaseURI',
