+++ packages/react-nft-collections/src/hooks/extensions/metadata/usePlaceholderUriUpdater.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { BytesLike } from '@ethersproject/bytes';
 import { Provider } from '@ethersproject/providers';
 import { Signer } from 'ethers';
@@ -15,10 +15,14 @@ export const usePlaceholderUriUpdater = ({
   contractAddress,
   signerOrProvider,
 }: Config) => {
-  return useContractWriteAndWait<[BytesLike]>({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn:
       'collections/ERC721/extensions/ERC721PrefixedMetadataExtension',
+  });
+
+  return useContractWriteAndWait<[BytesLike]>({
+    contractInterface,
     contractAddress,
     signerOrProvider,
     functionName: 'setPlaceholderURI',
