+++ packages/react-nft-collections/src/hooks/extensions/minting/useOneOfOneMinter.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { Provider } from '@ethersproject/providers';
 import { BigNumberish, BytesLike, Signer } from 'ethers';
 
@@ -28,9 +28,13 @@ export const useOneOfOneMinter = ({
   tokenURIs,
   access = 'ByOwner',
 }: Config) => {
-  return useContractWriteAndWait<ArgsType>({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn: 'collections/ERC721/extensions/ERC721OneOfOneMintExtension',
+  });
+
+  return useContractWriteAndWait<ArgsType>({
+    contractInterface,
     contractAddress,
     signerOrProvider,
     functionName:
