+++ packages/react-nft-collections/src/hooks/extensions/proceeds/useProceedsWithdraw.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { Provider } from '@ethersproject/providers';
 import { Signer } from 'ethers';
 
@@ -16,9 +16,13 @@ export const useProceedsWithdraw = ({
   contractAddress,
   signerOrProvider,
 }: Config) => {
-  return useContractWriteAndWait<ArgsType>({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn: 'collections/ERC721/extensions/ERC721SimpleProceedsExtension',
+  });
+
+  return useContractWriteAndWait<ArgsType>({
+    contractInterface,
     functionName: 'withdraw',
     contractAddress,
     signerOrProvider,
