+++ packages/react-nft-collections/src/hooks/extensions/sales/usePreSaleAllowlistMerkleRootUpdater.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { Provider } from '@ethersproject/providers';
 import { BytesLike, Signer } from 'ethers';
 
@@ -18,9 +18,13 @@ export const usePreSaleAllowlistMerkleRootUpdater = ({
   signerOrProvider,
   newRoot,
 }: Config) => {
-  return useContractWriteAndWait<ArgsType>({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn: 'collections/ERC721/extensions/ERC721PreSaleExtension',
+  });
+
+  return useContractWriteAndWait<ArgsType>({
+    contractInterface,
     functionName: 'setAllowlistMerkleRoot',
     contractAddress,
     signerOrProvider,
