+++ packages/react-nft-collections/src/hooks/extensions/sales/usePreSaleMinter.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { Provider } from '@ethersproject/providers';
 import { BigNumber, BigNumberish, BytesLike, Signer } from 'ethers';
 
@@ -31,9 +31,13 @@ export const usePreSaleMinter = ({
     contractAddress,
   });
 
-  const result = useContractWriteAndWait<ArgsType>({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn: 'collections/ERC721/extensions/ERC721PreSaleExtension',
+  });
+
+  const result = useContractWriteAndWait<ArgsType>({
+    contractInterface,
     functionName: 'mintPreSale',
     contractAddress,
     signerOrProvider,
