+++ packages/react-nft-collections/src/hooks/extensions/sales/usePublicSaleStatusToggler.ts
@@ -1,5 +1,5 @@
 import { ContractVersion } from '@0xflair/contracts-registry';
-import { useContractWriteAndWait } from '@0xflair/react-common';
+import { useContractAbi, useContractWriteAndWait } from '@0xflair/react-common';
 import { Provider } from '@ethersproject/providers';
 import { Signer } from 'ethers';
 
@@ -17,9 +17,13 @@ export const usePublicSaleStatusToggler = ({
   signerOrProvider,
   newValue,
 }: Config) => {
-  return useContractWriteAndWait<ArgsType>({
+  const contractInterface = useContractAbi({
     contractVersion,
     contractFqn: 'collections/ERC721/extensions/ERC721PublicSaleExtension',
+  });
+
+  return useContractWriteAndWait<ArgsType>({
+    contractInterface,
     functionName: 'togglePublicSaleStatus',
     contractAddress,
     signerOrProvider,
